requirements.txt
fastapi
uvicorn
streamlit
groq
pydantic
pandas
numpy
python-multipart
httpx
requests
python-dotenv
scikit-learn
sentence-transformers
transformers
torch
huggingface-hub
